% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetConsumerComplaints.R
\name{GetConsumerComplaints}
\alias{GetConsumerComplaints}
\title{Access All Consumer Finacial Protection Bureau}
\usage{
GetConsumerComplaints(before = NULL, after = NULL, limit = 20)
}
\arguments{
\item{before}{search for a near match based on string.}

\item{after}{A two letter name.}

\item{limit}{Number of entries returned. Default at 20 entries.}
}
\description{
This function allows the users to access Consumer Financial Protection Bureau API
through a variety of different parameters. Parameters should be left null if it is not of interest
so as to include all fields that have that parameter.
}
\details{
Returns a dataframe with 14 columns filtered based on search query.
}
\examples{
data <- GetConsumerComplaints()

data <- GetConsumerComplaints(limit = 25)
}
